Squashed 'externals/nitro/' changes from 50750b9..8683835
8683835 Merge pull request #13 from mdaus/support_static_tres
1634d7a Added treHandlerExists()
517c14c Added nitf_PluginRegistry_TREHandlerExists() function to test if a TRE handler exists.  Various cleanup on debug logging support.
9ddcd49 Minor style stuff - catch exception by const ref, catch ..., put macro in unnamed namespace
ae37697 Building test_static_plugin.cpp when static TRE support is turned on
697def4 Resurrecting support for static TRE linkage.  Now adding a configure option for this so it doesn't happen by default.  Also, each TRE is compiled into its own library with its own target rather than there being one library that has all of them
ff008dd Sending options() and configure() down a level.  Removing now unneeded convenience targets for the tests.  test_static_plugin.c doesn't actually exist so removing from TEST_FILTER
7fadf94 Updating maintainer email address
eee8c88 Commenting out variable name to avoid Windows warnings about unused variable
00838ee Updating static TRE macro based on TRE API change
e30e59b Merge commit 'c2cfdfa1e051f75857d7d9014f99d2e7cd098116'
c2cfdfa Squashed 'externals/coda-oss/' changes from da6141c..821312e
225ead1 Merge commit 'eca06504e370f2542b63d9b187cec7d97d261423'
eca0650 Squashed 'externals/coda-oss/' changes from 070dda5..da6141c

git-subtree-dir: externals/nitro
git-subtree-split: 86838355bc820bd10c9f44de95526fa5275eb034

+++ externals/coda-oss/modules/c++/math.poly/include/math/poly/TwoD.hpp
@@ -241,9 +241,9 @@ TwoD<_T>&
 TwoD<_T>::operator *= (const TwoD<_T>& p) 
 {
     TwoD<_T> tmp(orderX() + p.orderX(), orderY() + p.orderY());
-    for (int i = 0, xsz = mCoef.size() ; i < xsz; i++)
+    for (size_t i = 0, xsz = mCoef.size(); i < xsz; i++)
     {
-        for (int j = 0, ysz = p.mCoef.size() ; j < ysz; j++)
+        for (size_t j = 0, ysz = p.mCoef.size(); j < ysz; j++)
         {
             tmp.mCoef[i + j] += mCoef[i] * p.mCoef[j];
         }
